LPD-68467 Move ContentPreview component to common folder
